Multiple bug fixes and logging improvements

+++ Src/AngryWasp.Net/Config.cs
@@ -3,7 +3,7 @@ namespace AngryWasp.Net
     public class Node
     {
         public string Host { get; set; } = "127.0.0.1";
-        public int Port { get; set; } = Config.DEFAULT_PORT;
+        public ushort Port { get; set; } = Config.DEFAULT_PORT;
 
         public ulong PeerID { get; set; } = 0;
 
@@ -15,7 +15,7 @@ namespace AngryWasp.Net
 
     public class Config
     {
-        public const int DEFAULT_PORT = 3500;
+        public const ushort DEFAULT_PORT = 3500;
         public const int READ_BUFFER_SIZE = ushort.MaxValue;
         
         //The number of data failures to accept before diching the connection
